data_fetcher simulates feedback from OBD. testing.py hosts all our whitebox procedures
data_fetcher.py contains the function that will simulate some feedback from a car. This function returns a random gear and speed. This can be processed by Jack to give feedback to the user.

testing.py contains a preposition for how whitebox testing could be handled. This file contains a class 'Tests' which contains a test for each function in our  control system. It also contains a subclass 'fullTest' which can be used to cycle through every test from 'Tests' in a single go.